WIP: support colornames
