+++ setup.py
@@ -80,7 +80,7 @@ if __name__ == '__main__':
         maintainer='Timo Furrer',
         maintainer_email='tuxtimo@gmail.com',
         include_package_data=True,
-        package_data={'': ['colorful/data/*.txt', 'README.md']},
+        package_data={'': ['colorful/data/*.txt', 'colorful/data/*.json', 'README.md']},
         packages=find_packages(exclude=['*tests*']),
         install_requires=requirements,
         classifiers=[
